Optimize System Loading for Vite Chunking (#7104)
* optimize system loading for vite chunking

* fix tsc errors
