+++ packages/client-core/src/admin/adminRoutes.tsx
@@ -1,9 +1,9 @@
 import React, { Suspense, useEffect } from 'react'
 import { Redirect, Switch } from 'react-router-dom'
 
-import { Engine } from '@xrengine/engine/src/ecs/classes/Engine'
 import { useEngineState } from '@xrengine/engine/src/ecs/classes/EngineState'
-import { initializeCoreSystems, initializeSceneSystems } from '@xrengine/engine/src/initializeEngine'
+import { initializeCoreSystems } from '@xrengine/engine/src/initializeCoreSystems'
+import { initializeSceneSystems } from '@xrengine/engine/src/initializeSceneSystems'
 
 import CircularProgress from '@mui/material/CircularProgress'
