+++ packages/engine/src/scene/functions/loaders/ParticleEmitterFunctions.test.ts
@@ -2,6 +2,8 @@ import assert from 'assert'
 import { afterEach, beforeEach, describe } from 'mocha'
 import { createSandbox, SinonSandbox } from 'sinon'
 
+import { initializeCoreSystems } from '@xrengine/engine/src/initializeCoreSystems'
+
 import { Engine } from '../../../ecs/classes/Engine'
 import { Entity } from '../../../ecs/classes/Entity'
 import { World } from '../../../ecs/classes/World'
@@ -10,7 +12,7 @@ import { addEntityNodeChild, EntityTreeNode } from '../../../ecs/functions/Entit
 import { createEntityNode } from '../../../ecs/functions/EntityTree'
 import { initSystems } from '../../../ecs/functions/SystemFunctions'
 import { SystemUpdateType } from '../../../ecs/functions/SystemUpdateType'
-import { createEngine, initializeCoreSystems, setupEngineActionSystems } from '../../../initializeEngine'
+import { createEngine, setupEngineActionSystems } from '../../../initializeEngine'
 
 describe('ParticleEmitterFunctions', async () => {
   let entity: Entity
