+++ packages/engine/src/xr/XRSystem.ts
@@ -1,6 +1,14 @@
 import { createActionQueue, getState, removeActionQueue } from '@xrengine/hyperflux'
 
+import XR8 from './8thwall/XR8'
 import { World } from './../ecs/classes/World'
+import XRAnchorSystem from './XRAnchorSystem'
+import XRCameraSystem from './XRCameraSystem'
+import XRControllerSystem from './XRControllerSystem'
+import XRDepthOcclusion from './XRDepthOcclusion'
+import XRHapticsSystem from './XRHapticsSystem'
+import XRLightProbeSystem from './XRLightProbeSystem'
+import XRScenePlacementShader from './XRScenePlacementShader'
 import { endXRSession, requestXRSession, xrSessionChanged } from './XRSessionFunctions'
 import { XRAction, XRState } from './XRState'
 
@@ -46,14 +54,14 @@ export default async function XRSystem(world: World) {
     execute,
     cleanup,
     subsystems: [
-      () => import('./8thwall/XR8'),
-      () => import('./XRAnchorSystem'),
-      () => import('./XRCameraSystem'),
-      () => import('./XRControllerSystem'),
-      () => import('./XRHapticsSystem'),
-      () => import('./XRLightProbeSystem'),
-      // () => import('./XRDepthOcclusion'),
-      () => import('./XRScenePlacementShader')
+      () => Promise.resolve({ default: XR8 }),
+      () => Promise.resolve({ default: XRAnchorSystem }),
+      () => Promise.resolve({ default: XRCameraSystem }),
+      () => Promise.resolve({ default: XRControllerSystem }),
+      () => Promise.resolve({ default: XRHapticsSystem }),
+      () => Promise.resolve({ default: XRLightProbeSystem }),
+      // () => Promise.resolve({ default: XRDepthOcclusion }),
+      () => Promise.resolve({ default: XRScenePlacementShader })
     ]
   }
 }
